Scripts/requirements.txt
cymysql
pillow
requests
requests-toolbelt